requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.ini
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
charset-normalizer==3.3.2
    # via httpx
click==8.1.7
    # via uvicorn
fastapi==0.112.0
    # via -r requirements.ini
greenlet==3.0.3
    # via -r requirements.ini
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.14.7
    # via httpx
httpx==0.21.3
    # via -r requirements.ini
idna==3.7
    # via
    #   anyio
    #   rfc3986
psycopg==3.2.1
    # via -r requirements.ini
pydantic==2.8.2
    # via
    #   -r requirements.ini
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
rfc3986==1.5.0
    # via httpx
sniffio==1.3.1
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.31
    # via -r requirements.ini
starlette==0.37.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.30.5
    # via -r requirements.ini

pyjwt==2.8.0
pytz==2024.1